refactor: decompose studio.service.ts into 4 modules (Sprint 051)
Split 1137 LOC monolith into focused modules:
- studio-sections.service.ts: section detection, parsing, replacement
- studio-stems.service.ts: source track, stems, transcription, MIDI
- studio-operations.service.ts: version mgmt, track ops, mashup/persona
- studio-realtime.service.ts: telegram, generation, realtime subscriptions

Main studio.service.ts re-exports for backward compatibility.
