+++ src/services/studio/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Studio Service — decomposed modules
+ * Sprint 051: split from studio.service.ts (1137 LOC → 4 modules)
+ */
+export * from "./studio-sections.service";
+export * from "./studio-stems.service";
+export * from "./studio-operations.service";
+export * from "./studio-realtime.service";
